move chainsubscriber to api
