+++ packages/api/src/app.ts
@@ -1,6 +1,7 @@
 import 'module-alias/register';
 import { utils, config, database } from '@impactmarket/core';
 import express from 'express';
+import job from './subscriber/job';
 
 import serverLoader from './server';
 
@@ -23,6 +24,9 @@ export async function startServer() {
 
     await serverLoader(app);
     utils.Logger.info('📡 Express server loaded');
+
+    job();
+    utils.Logger.info('⏱️ Chain Subscriber starting');
 
     return app.listen(config.port, () => {
         utils.Logger.info(`
